Refactor LoanDAO and Error handling: update query to fetch loans, introduce ErrorType enum for standardized error responses, and enhance LoanService validation

+++ src/main/java/com/example/biblioteca/service/LoanService.java
@@ -6,8 +6,6 @@ import com.example.biblioteca.model.Loan;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 
@@ -19,32 +17,83 @@ public class LoanService {
 
   @Transactional
   public void createLoan(Loan loan) {
+    if (loan.getUser() == null || loan.getUser().isEmpty()) {
+      throw new IllegalArgumentException("Nome do usuário não pode ser nulo ou vazio");
+    }
+
+    if (loan.getLoanDate() == null) {
+      throw new IllegalArgumentException("Data de retirada não pode ser nula");
+    }
+
+    if (loan.getReturnDate() == null) {
+      throw new IllegalArgumentException("Data de retorno não pode ser nula");
+    }
+
     Set<Book> books = loan.getBooks();
-    System.out.println(books);
-    System.out.println(loan);
     if (books == null) {
       throw new IllegalArgumentException("Books cannot be null");
     }
+
     loanDAO.create(loan);
   }
 
   @Transactional
   public Loan findLoanById(Long id) {
-    return loanDAO.findById(id);
+    Loan loan = loanDAO.findById(id);
+    if (loan == null) {
+      throw new IllegalArgumentException("Empréstimo não encontrado");
+    }
+    return loan;
   }
 
   @Transactional
   public List<Loan> findAllLoans() {
-    return loanDAO.findAll();
+    List<Loan> loan = loanDAO.findAll();
+
+    if (loan == null) {
+      throw new IllegalArgumentException("Empréstimos não encontrados");
+    }
+    return loan;
   }
 
   @Transactional
-  public void updateLoan(Loan loan) {
-    loanDAO.update(loan);
+  public void updateLoan(Long id, Loan updatedLoan) {
+    Loan existingLoan = loanDAO.findById(id);
+    if (existingLoan == null) {
+      throw new IllegalArgumentException("Empréstimo não encontrado");
+    }
+
+    if (updatedLoan.getUser() == null || updatedLoan.getUser().isEmpty()) {
+      throw new IllegalArgumentException("Nome do usuário não pode ser nulo ou vazio");
+    }
+
+    if (updatedLoan.getLoanDate() == null) {
+      throw new IllegalArgumentException("Data de retirada não pode ser nula");
+    }
+
+    if (updatedLoan.getReturnDate() == null) {
+      throw new IllegalArgumentException("Data de retorno não pode ser nula");
+    }
+
+    Set<Book> books = updatedLoan.getBooks();
+    if (books == null) {
+      throw new IllegalArgumentException("Books cannot be null");
+    }
+
+    existingLoan.setUser(updatedLoan.getUser());
+    existingLoan.setLoanDate(updatedLoan.getLoanDate());
+    existingLoan.setReturnDate(updatedLoan.getReturnDate());
+    existingLoan.setBooks(updatedLoan.getBooks());
+
+    loanDAO.update(existingLoan);
   }
 
   @Transactional
   public void deleteLoan(Long id) {
+    Loan existingLoan = loanDAO.findById(id);
+    if (existingLoan == null) {
+      throw new IllegalArgumentException("Empréstimo não encontrado");
+    }
     loanDAO.delete(id);
   }
 }
